requirements.txt
absl-py==0.9.0
asgiref==3.2.7
astor==0.8.1
cycler==0.10.0
Django==3.0.6
gast==0.3.3
grpcio==1.28.1
h5py==2.10.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
Markdown==3.2.1
matplotlib==3.2.1
mock==4.0.2
numpy==1.18.4
Pillow==7.1.2
pkg-resources==0.0.0
protobuf==3.11.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
Werkzeug==1.0.1
